requirements.txt
flask
flask-sqlalchemy 
flask-migrate
openai 
python-dotenv 
pdfplumber